Move getSprintPeriod to SprintPlanningService
